Added new group Model
