+++ UIMentoringSession/MenteesChallenges/Pets Main Card UI Chellenge/OleksandrKharchenko/Model/Olkha_PetsModels.swift
@@ -4,90 +4,69 @@ import UIKit
 struct OlKha_PetsViewModel {
     let name: String
     let age: String
-    let breadAnimal: String
+    let breedAnimal: String
     let type: AnimalsTypes
     let animal: UIImage
 }
-
-struct OlKha_FilterViewModel {
-    let title: String
-    let type: AnimalsTypes
-    let animal: UIImage
-}
-
-let itemsFilter: [OlKha_FilterViewModel] = [
-    .init(title: "All",
-          type: .allPets,
-          animal: RImage.allPet()!
-         ),
-    .init(title: "Cat",
-          type: .cats,
-          animal: RImage.catPet()!
-         ),
-    .init(title: "Dog",
-          type: .dogs,
-          animal: RImage.dogPet()!
-         ),
-]
 
 let itemsAnimals: [OlKha_PetsViewModel] = [
     .init(name: "Gary",
           age: "3 years",
-          breadAnimal: "Yorkshire Terrier",
+          breedAnimal: "Yorkshire Terrier",
           type: .dogs,
           animal: RImage.garyDog()!
          ),
     .init(name: "Peach",
           age: "2.5 years",
-          breadAnimal: "Half-breed",
+          breedAnimal: "Half-breed",
           type: .cats,
           animal: RImage.peachCat()!
          ),
     .init(name: "Whitney",
           age: "2 months",
-          breadAnimal: "British Longhair",
+          breedAnimal: "British Longhair",
           type: .cats,
           animal: RImage.whitheyCat()!
          ),
     .init(name: "Buggy",
           age: "4 months",
-          breadAnimal: "Jack Russell Terrier",
+          breedAnimal: "Jack Russell Terrier",
           type: .dogs,
           animal: RImage.buggyDog()!
          ),
     .init(name: "Willie",
           age: "1.5 years",
-          breadAnimal: "Samoyed",
+          breedAnimal: "Samoyed",
           type: .dogs,
           animal: RImage.willieDog()!
          ),
     .init(name: "Kiwi",
           age: "1 years",
-          breadAnimal: "Yorkshire Terrier",
+          breedAnimal: "Yorkshire Terrier",
           type: .dogs,
           animal: RImage.kiwiDog()!
          ),
     .init(name: "Stitch",
           age: "1 year",
-          breadAnimal: "European cat",
+          breedAnimal: "European cat",
           type: .cats,
           animal: RImage.stitchCat()!
          ),
     .init(name: "Cake",
           age: "2 months",
-          breadAnimal: "Welsh Corgi",
+          breedAnimal: "Welsh Corgi",
           type: .dogs,
           animal: RImage.cakeDog()!
          ),
     .init(name: "Moon",
           age: "2.5 years",
-          breadAnimal: "Siberian cat",
+          breedAnimal: "Siberian cat",
           type: .cats,
           animal: RImage.moonCat()!
          ),
     .init(name: "Spike",
           age: "1.5 years",
-          breadAnimal: "Maine Coon",
+          breedAnimal: "Maine Coon",
           type: .cats,
           animal: RImage.spikeCat()!
          ),
